Fixed issue with package version getter
