+++ tests/test_utils/test_version_utils.py
@@ -2,6 +2,7 @@ import os
 import sys
 import logging
 import unittest
+from unittest.mock import patch, MagicMock
 
 # Logging Setup
 logging.basicConfig()
@@ -74,3 +75,24 @@ class TestVersionUtils(unittest.TestCase):
         result = version_utils.compare_versions(version_a="2.2.3", version_b="1.6.7")
         self.assertEqual(expected, result)
 
+    @patch('os.path.exists')
+    def test_get_package_version_bad_path(self, mock_eval):
+        result = version_utils.get_package_version(package_path="mocked_package_path")
+        mock_eval.assert_called_once()
+        expected = "0.0.0"
+        self.assertEqual(expected, result)
+
+    @patch('sys.path')
+    @patch('os.path.exists')
+    def test_get_package_version(self, mock_exists, mock_path):
+        mock_exists.return_value = True
+        mock_path.return_value = ['/mocked/path', 'mocked_package_path']
+        with patch('builtins.__import__') as mock_import:
+            mock_instance = MagicMock()
+            mock_instance.__version__ = '1.2.3'
+            mock_import.return_value = mock_instance
+            result = version_utils.get_package_version(package_path="mocked_package_path")
+            mock_exists.assert_called_once()
+            mock_import.assert_called_once()
+            expected = '1.2.3'
+            self.assertEqual(expected, result)
